fix: add cross-platform support for Windows builds
This commit adds platform-specific implementations for Unix-specific
syscall operations to enable Windows cross-compilation and native builds.

Changes:
- pkg/audit: Split checkDiskSpace into audit_unix.go and audit_windows.go
- pkg/vault: Split CheckDiskSpace into vault_unix.go and vault_windows.go
- internal/mcp: Split policy loading helpers into policy_unix.go and policy_windows.go
- cmd/secretctl: Split signal handling and core dump functions into
  run_unix.go and run_windows.go

Additional fixes:
- Updated release.yml to use libwebkit2gtk-4.1-dev for Ubuntu 24.04
- Added webkit2_41 build tag for Wails on Ubuntu 24.04
