+++ pkg/audit/audit.go
@@ -39,7 +39,6 @@ import (
 	"path/filepath"
 	"runtime"
 	"sync"
-	"syscall"
 	"time"
 
 	"golang.org/x/crypto/hkdf"
@@ -912,24 +911,6 @@ func (l *Logger) rewriteLogFile(path string, events []AuditEvent) error {
 	return os.Rename(tempPath, path)
 }
 
-// checkDiskSpace verifies sufficient disk space for audit log writes
-func (l *Logger) checkDiskSpace() error {
-	var stat syscall.Statfs_t
-	if err := syscall.Statfs(l.path, &stat); err != nil {
-		// If audit directory doesn't exist yet, check parent
-		parentDir := filepath.Dir(l.path)
-		if err := syscall.Statfs(parentDir, &stat); err != nil {
-			// Log warning but don't block audit operation
-			fmt.Fprintf(os.Stderr, "warning: failed to check disk space for audit: %v\n", err)
-			return nil
-		}
-	}
-
-	available := stat.Bavail * uint64(stat.Bsize)
-	if available < MinAuditDiskSpace {
-		return fmt.Errorf("audit: insufficient disk space: only %d bytes available, need at least %d",
-			available, MinAuditDiskSpace)
-	}
-
-	return nil
-}
+// checkDiskSpace is defined in platform-specific files:
+// - audit_unix.go: Unix/Linux/macOS implementation using syscall.Statfs
+// - audit_windows.go: Windows stub (returns nil)
